refactored price per hour to type Integer

+++ server/app/models.py
@@ -161,7 +161,7 @@ class Space(db.Model, SerializerMixin):
     owner_id = db.Column(db.Integer, db.ForeignKey('users.id', ondelete="CASCADE"), nullable=False)
     title = db.Column(db.String, nullable=False)
     description = db.Column(db.String, nullable=True)
-    price_per_hour = db.Column(db.Numeric(10, 2), nullable=False)
+    price_per_hour = db.Column(db.Integer, nullable=False)
     status = db.Column(db.String, nullable=False, default='available')
     images = db.Column(db.ARRAY(db.Text), default=[], nullable=True)
     space_type = db.Column(db.String, nullable=False)
